+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## Release 4.9
 
 * #513 AgEntityOverly with a single root resolver doesn't update the entity
+* #514 A simpler way to redefine root data resolver
 
 ## Release 4.8